// ccr-ui/src/types/generated/environment/EnvironmentInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EnvironmentType } from "./EnvironmentType";

/**
 * 环境概要信息（用于前端展示）
 */
export type EnvironmentInfo = { id: string, env_type: EnvironmentType, name: string, display_name: string, description: string, is_active: boolean, };